test(workbench): enable injection of workbench services in unit tests
The following services could not be injected in unit tests:
- `WorkbenchDialogService`
- `WorkbenchMessageBoxService`
- `WorkbenchPopupService`
