+++ projects/scion/workbench/src/lib/message-box/workbench-message-box-service.provider.ts
@@ -1,0 +1,23 @@
+/*
+ * Copyright (c) 2018-2026 Swiss Federal Railways
+ *
+ * This program and the accompanying materials are made
+ * available under the terms of the Eclipse Public License 2.0
+ * which is available at https://www.eclipse.org/legal/epl-2.0/
+ *
+ * SPDX-License-Identifier: EPL-2.0
+ */
+
+import {Provider} from '@angular/core';
+import {ɵWorkbenchMessageBoxService} from './ɵworkbench-message-box.service';
+import {WorkbenchMessageBoxService} from './workbench-message-box.service';
+
+/**
+ * Provides {@link WorkbenchMessageBoxService} for dependency injection.
+ */
+export function provideWorkbenchMessageBoxService(): Provider[] {
+  return [
+    ɵWorkbenchMessageBoxService,
+    {provide: WorkbenchMessageBoxService, useExisting: ɵWorkbenchMessageBoxService},
+  ];
+}
