+++ projects/scion/workbench/src/lib/part/workbench-part-context.provider.ts
@@ -9,9 +9,9 @@
  */
 
 import {InjectionToken, Provider} from '@angular/core';
-import {provideWorkbenchDialogService} from '../dialog/ɵworkbench-dialog.service';
-import {provideWorkbenchMessageBoxService} from '../message-box/ɵworkbench-message-box.service';
-import {provideWorkbenchPopupService} from '../popup/ɵworkbench-popup.service';
+import {provideWorkbenchMessageBoxService} from '../message-box/workbench-message-box-service.provider';
+import {provideWorkbenchDialogService} from '../dialog/workbench-dialog-service.provider';
+import {provideWorkbenchPopupService} from '../popup/workbench-popup-service.provider';
 
 /**
  * DI token to register providers available for DI if in the context of a workbench part.
