+++ projects/scion/workbench/src/lib/testing/workbench.provider.ts
@@ -16,9 +16,6 @@ import {MAIN_AREA_INITIAL_PART_ID} from '../layout/ɵworkbench-layout';
 import {ComponentFixtureAutoDetect} from '@angular/core/testing';
 import {Router} from '@angular/router';
 import {PartId} from '../workbench.identifiers';
-import {provideWorkbenchDialogService} from '../dialog/ɵworkbench-dialog.service';
-import {provideWorkbenchMessageBoxService} from '../message-box/ɵworkbench-message-box.service';
-import {provideWorkbenchPopupService} from '../popup/ɵworkbench-popup.service';
 
 /**
  * Enables and configures the SCION Workbench in tests., returning a set of dependency-injection providers to be registered in Angular.
@@ -50,8 +47,5 @@ export function provideWorkbenchForTest(config?: WorkbenchConfig & {mainAreaInit
     provideEnvironmentInitializer(() => inject(Router).initialNavigation()),
     provideEnvironmentInitializer(() => localStorage.clear()),
     provideEnvironmentInitializer(() => window.name = ''),
-    provideWorkbenchDialogService(),
-    provideWorkbenchMessageBoxService(),
-    provideWorkbenchPopupService(),
   ]);
 }
